Caching: Add `Node.CACHE_VERSION` to node attributes
Recently, the hashing mechanism was changed to remove all version
information, both `aiida-core` and the plugin if applicable, from the
hash computation. The reason was that with that information, each
version change in core and plugin would essentially reset the cache.
By removing the version information, updating the code would no longer
invalidate all the cache.

The downside of this change, however, is that when the implementation of
a plugin, most notably a `CalcJob` plugin, changes significantly such
that a change in its attributes/inputs would really change its content,
this would no longer be reflected in the hash, which would remain
identical. This could lead to false positives when users update certain
plugins.

To give some control to plugin developers in case of changes where the
hash would have to be effectively reset, the `Node` class now defines
the `CACHE_VERSION` class attribute. By default this is set to `None`
but a `Data` plugin can change it to some string value. To facilitate
the same API for process plugins, such as `CalcJob`'s, the
`CACHE_VERSION` attribute is also added to the `Process` base class.
This value is automatically added to the node's attributes, under the
key `_aiida_cache_version` such that is is included in the computed hash.
This way, a plugin developer can change this attribute in their plugin
to force the hash to change.

The version has to be stored in the attributes of the node and cannot
just be taken from the plugin's class attribute directly when computing
the hash, because in that case when the hash is recalculated in the
future, a potentially different value of `CACHE_VERSION` could be used
(whichever version the plugin has that is installed at that time).