Config: Refactor the `create_profile` function and method
The `aiida.manage.configuration.create_profile` function and the
`Config.create_profile` method that it invokes are refactored to allow
defining the broker plugin and configuration. This is useful now that
the broker for a profile can be configured to be other than the default
RabbitMQ implementation. Most notably it can be set to `None` to run
AiiDA in a mode without a broker.