`BaseRestartWorkChain`: Fix handler overrides used only first iteration (#6308)
The handler overrides were only being used in the first iteration of the
`BaseRestartWorkChain`. The code was using `pop` to obtain the `enabled`
and `priority` attributes, thereby removing them from the overrides that
are stored in the context. In the following iteration, the original
values wouldn't be there anymore. The fix is to simply replace `pop`
with `get`.