CLI: Allow `verdi computer delete` to delete associated nodes  (#6334)
The command would previously except if the computer had associated nodes
which was pretty much always going to be the case, making the command
pretty useless.

The command now checks for associated nodes and prompts the user to
confirm to delete it. After confirmation the nodes are deleted, followed
by the computer itself. A `--dry-run` option is added to run the command
in dry mode, so the user can see which nodes would be deleted.