`PsqlDosBackend`: Add missing `repository_uri` to the model (#6342)
The field not being declared means an option for it was not added to the
dynamically generated `verdi profile setup core.psql_dos`. This would
cause the command to fail as the initialization of the storage would
raise an error complaining about the missing `repository_uri`. This was
missed because the `core.psql_dos` storage was excluded from the
parametrized test for the `verdi profile setup` command but this is now
added.