Config: Prioritize CWD over `AIIDA_PATH` in determining config dir
The current order of priority would have `AIIDA_PATH` always override
the current working directory hierarchy in determining the location of
the configuration directory. This would essentially render the CWD logic
pointless as soon as `AIIDA_PATH` is defined, which is the case for many
existing installations as that was the only way up till now to control
the location of the configuration directory. It also breaks the
functionality of `verdi init` as the created directory won't be picked
up.

Therefore, the order is switched and the CWD takes precedence. The
default config directory location, which is currently the user's home
folder, is hereby ignored when going up the hierarchy. This is because
this folder is likely to almost always exist and be present in the CWD
hierarchy and would therefore always override the `AIIDA_PATH`.