CLI: Optional sorting for `verdi code export`
When running `verdi code export`, the generated yaml file is
automatically sorted. However, the resulting alphabetical ordering is
not what is typically found in the code yaml files in the
`aiida-code-registry` and `aiida-resource-registry`. This PR turns the
sorting off by default, and provides a command-line option to turn on
the sorting.

If we don't find any other use cases for this flag, I'd also be fine to
just turn off the yaml sorting alltogether, as I don't think this is
the format that we'd usually want in the yaml file?

I found the previous behavior to be a slight annoyance, while setting up
some unfamiliar codes by exporting them from an existing archive and
modifying the resulting yaml files accordingly.