Config: Revert checking of CWD for config directory location
In 1059b5f2d365e0f2ea4dea4d3c9343ce77829cfe the logic to determine the
location of the configuration directory was modified to also check the
hierarchy of the current working directory for an existing directory.
After discussion, it seems that this change is more impactful than
anticipated and needs to be discussed in more depth before being
released. Therefore the change is reverted for now.