Alternative solution to `is_backup_implemented`
The `is_backup_implemented` property was added to the storage interface
such that the base class can call this before calling the actual storage
backend. The reason is that the creation of the output folder has to be
done before `_backup_backend` is called, but if that then raises
`NotImplementedError` the output directory would have been created for
nothing and remain empty.

The current solution adds an additional method to the abstract interface
called `is_backup_implemented`. Storage plugins should override this to
return `True` in addition to implementing `_backup_backend`. This feels
a bit redundant and can easily be forgotten by developers.

Here as an alternative, the `is_backup_implemented` is removed and the
base class `backend` will simply create the output folder and then
delete it if the plugin raises `NotImplementedError`.