Storage: Add backup mechanism to the interface (#6069)
Up until now, backing up the data of a profile had to be performed
manually by the user. The `StorageBackend` abstract class for storage
plugins adds the scaffolding to automate this procedure.

The class adds the `backup` method that takes as arguments the target
directory where to create the backup and the number of backups to keep.
The base class already implements the base functionality of creating the
target directory and stores a copy of the global configuration file,
removing all profiles but the one whose data is being backed up. The
command then calls the `_backup` method, which is what should be
implemented by the storage plugins.

The base class uses the `backup_utils.BackupManager` utility of the
`disk_objectstore` package, which provides functionality to support
backing up data directly over SSH. It also contains methods to create
incremental backups efficiently using `rsync` and hard linking. For now
this behavior is not enforced onto storage plugins so they can opt to
not use this functionality should they choose to. For now, only the
`core.psql_dos` plugin implements the backup mechanism. For backing up
the PostgreSQL data it relies on the `pg_dump` executable and the
repository is delegated to the functionality implemented in the
`disk_objectstore` package itself.

The new functionality is exposed on the command line interface through
the command `verdi storage backup`.