Add first version of `--flat` option.
This option allows dumping of a CalcJob (or of a simple WorkChain that
only calls a single CalcJob) in a flat directory, without creating the
hierarchy of `raw_inputs`, `raw_outputs`, and `node_inputs`. This might
be useful for cases where AiiDA is only used to run a calculation and
dump the results in a specific custom path, where the custom path is
dictated by the other code that calls AiiDA to submit calculations
(e.g. what we are currently working on aiida-koopmans).