Devops: Use `uv` to install package in CI and CD (#6363)
The `uv` utility is a drop in replacement for `pip` by the creators
of `ruff`` that comes with a significant speedup. All installation of
`aiida-core` in the CI/CD workflows are replaced with `uv` except
for on in the `test-install.yml` workflow, just to have some check that
`pip` keeps working (although it is unlikely to ever break).