test
Changed `--flat` option to still create subdirectories for the
individual steps of the WorkChain. Instead, just the subdirectories
per CalcJob are removed.

Generalized the dumping of outputs that it doesn't only dump
`retrieved` -> With this, it's dumping a whole range of aiida nodes,
basically all the parsed outputs, which are mainly numpy arrays dumped
as `.npy` files. Add an option to enable this, as it might not be
necessary to dump all of those. Currently, I just defined a global
variable in the file, but this will eventually become a class attribute
of the ProcessDumper class.