♻️ First working OOP-version using `ProcessDumper`
To avoid having to pass all the arguments
`include_node_inputs`, `include_attributes/extras`, `overwrite`, `flat`,
`all_aiida_nodes` through the different functions, everything related to
the dumping is now compiled in the `ProcessDumper` class, which defines
the main entry-point method `dump`. For nested workflows, this is
recursively called (as before). Once `CalculationFunction` nodes are
reached, their content is dumped via `dump_calculation_node`. The
helper functions to create and validate labels and paths of nested
subdirectories are also methods of the `ProcessDumper`.

Introduced the `parent_process` class attribute which is dynamically
generated from the parent_node, and which is used to generate the main
README, which is only created when the dumping is done via the `verdi`
CLI. For the other functions, this concept does not make sense, due to
the recursion, so the respective `process_node`s (which are changing
during the recursion) are always passed as arguments.

Next steps:
- Update tests to actually test the new implementations
- Update docstrings
- Add section to `How to work with data` section of the docs
- If the `OverridableOptions` are only used here, they can also just be
  defined as normal `click` options (however, we can also start thinking
  about the `verdi archive dump` functionality that we should start
  implementing soon)