CLI: Add the command `verdi presto` (#6351)
This command aims making setting up a new profile as easy as possible.
It intentionally provides a limited amount of options to customize the
profile and by default does not require any options to be specified at
all. For full control, the command `verdi profile setup` should be used
instead.

The main goal for this command is to setup a profile with minimal
requirements to make it easy to install AiiDA and get started as quickly
as possible. Therefore, by default, the created profile uses the
`core.sqlite_dos` storage plugin which does not require any services,
such as PostgreSQL and RabbitMQ are not required. This _does_ mean,
however, that not all functionality of AiiDA is available, most notably
running the daemon and submitting processes to said daemon.

The command performs the following actions:

* Create a new profile that is set as the new default
* Create a default user for the profile (can be configured through options)
* Set up the localhost as a `Computer` and configure it
* Set a number of configuration options with sensible defaults

In the future, it may be possible to incorporate the functionality of
the `verdi quicksetup` command that automatically creates the role and
database in PostgreSQL necessary for a profile with the `core.psql_dos`
storage plugin. This would allow `verdi quicksetup` to be retired
leaving just `verdi presto` and `verdi profile setup` to provide all the
profile setup needs.