CI: Test with newer RabbitMQ versions
Version 3.6/3.7 are super old and unsupported,
let's test with newer, supported versions instead.

We still want to support v3.8, since that's the last one
that doesn't require any extra configuration (i.e. setting longer
timeouts), but this version is tested as part of normal tests in ci-code.yml

Since AiiDA interacts with RabbitMQ only via the kiwipy interface,
it's unlikely that changes to AiiDA would break it. Therefore, the
RabbitMQ tests are moved to the nightly run.