CLI: Fix dry-run resulting in critical error in `verdi archive import` (#6403)
The import module raises `ImportTestRun` after a dry-run of a importing
archive has finished successfully. This was not caught separately by
the `verdi archive import` command resulting in a critical error being
returned even in the case of a successful dry-run. The exception is now
caught and a success message is logged instead.