Devops: Move RabbitMQ CI to nightly and update versions (#6419)
The `rabbitmq.yml` was running a subset of the unit test suite against
a number of versions of RabbitMQ. The versions tested were very outdated
and are not officially supported anymore. The versions are updated to
3.11, 3.12, and 3.13.

Since the introduction, the tests have never really failed, which is not
too suprising as the code in `aiida-core` doesn't actually directly
interface with RabbitMQ. Rather this is done in `kiwipy`. Therefore, the
tests are moved to the nightly workflow such that they are no longer
run on every push and PR, which is overkill.