CLI: Ensure deprecation warnings are printed before any prompts
The `click.Command` has a `deprecated` attribute, which is a bool, that
when set to `True` results in a printed deprecation warning on usage,
and adaptions of the help page to indicate the command as deprecated.

The `click.option` prompts are however invoked before that warning is
printed. This can lead to users filling out all prompts before they are
finally greeted with the deprecation message. Therefore we customize
the `click.Command` class introducing `aiida.cmdline.groups.VerdiCommand`
to move the printing logic to before the first prompt is shown.

The `VerdiCommandGroup` sets `command_class` to this new custom class
such that all `verdi` commands automatically use it. To deprecate a
command, the `deprecated` argument now just has to be set to the desired
deprecation message.