CLI: Add automatic PostgreSQL management to `verdi presto` (#6432)
The functionality of `verdi quicksetup` that automatically creates a
PostgreSQL user and database was still relied upon by quite a number of
users. Therefore, this functionality is integrated into `verdi presto`
so that `verdi quicksetup` can finally be deprecated.

Since the crux of `verdi presto` is to be as automatic as possible and
require no prompting and as little configuration as necessary, only a
few options are exposed. To use PostgreSQL instead of SQLite, the
`--use-postgres` switch needs to be toggled. Only the hostname and port
of the PostgreSQL server can be configured, as well as the username and
password of an existing user with which to connect.