Engine: Set the `to_aiida_type` as default inport port serializer
The `to_aiida_type` serializer automatically converts a number of Python
base types to the corresponding `Data` node when passed as an input to a
process. Process functions already automatically set this serializer as
a default, which simplifies the life of users and developers.

Here, the `to_aiida_type` is now set as the default serializer for all
input ports. The only exception is if the port is a metadata input port,
in which case the data is stored directly in the attributes of the
process node and so should not be converted to a node. It is also
skipped if the port itself already declares a serializer.