CLI: Deprecate accepting `0` for `default_mpiprocs_per_machine`
The `verdi computer setup` command would accept the value `0` for the
`default_mpiprocs_per_machine` input. This was added as a value for
users to indicate as to _not_ set a value for the property. But the CLI
has the `!` marker for this purpose. The `ComputerBuilder` converts the
`0` to `None` but this conversion now raises a deprecation warning.