CLI: Accept mulitple node identifiers in `verdi node graph generate`
The `--identifier` option allows the user to specify which identifier to
label nodes in the graph with: pk, uuid or label. Here, the interface is
updated to allow specifying multiple identifiers, e.g.:

    verdi node graph generate --identifier pk uuid -- <PK>

If more than one identifier type is specified, the resulting identifiers
for each node are joined using a `|` character.