Devops: Allow unit test suite to be ran against SQLite
The unit test suite currently automatically creates a test profile using
the `core.psql_dos` storage plugin. The reason is mostly historical as
originally the `core.psql_dos` storage was the only. The downside of
this storage plugin, however, is that it requires a PostgreSQL database.
This typically requires a PostgreSQL service to be running, making it
more difficult to get a working development environment. This was made
slightly easier through the use of `pgtest`, which would create a PSQL
cluster on-the-fly, but this would still require postgres libraries to
be installed.

Nowadays, storage plugins are customizable and the `core.sqlite_dos`
replaces PostgreSQL with the serviceless SQLite. The dependencies for
these are a lot easier to install, making it perfect for testing.
However, the test suite automatically configures a `core.psql_dos`
profile and some tests are not compatible with SQLite since they
directly depend on functionality that is only supported by PSQL and not
SQLite.

In this commit, changes are made that allow running the test with
SQLite. Two new pytest markers are added to the test suite:

* `requires_pgsql`: for tests that are known to require PostgreSQL (i.e.
  do not work with SQLite)
* `presto`: an alias for `not requires_rmq and not requires_pgsql`. The
  tests are marked with `presto` automagically in
  `tests/conftest.py:pytest_collection_modifyitems`

A new CI job is added `test-presto` that runs the tests with the
`presto` marker.