Remove test for `make_aware` using fixed date
The test compares the timezones between a fixed date and of the current
date. This causes problem with time zones that use daylight saving time
(DST). Fixing the test makes the test more complicated that it is worth
it. The `make_aware` function is thoroughly enough testet but the
remaining tests.