Docs: Rework the installation section
The `verdi setup` and `verdi quicksetup` commands have been deprecated
and replaced by `verdi profile setup` and `verdi presto`. The
installation docs were heavily outdated and the flow was scattered.

The biggest change is that there now is a "quick install guide" that
relies on `verdi presto` to provide an install route that is fool proof
and will work on almost any system in a minimal amount of commands.

Then there is the "complete installation guide" that provides all the
details necessary to fully customize an installation.