Devops: Remove `verdi tui` from CLI reference documentation (#6464)
The `verdi-autodocs` pre-commit hook automatically generates the
reference documentation for `verdi`. It does so based on the available
commands. The `verdi tui` command is added dynamically if and only if
the optional dependency `trogon` is installed.

Since having the dependency installed slows down the loading time
`verdi` significantly, many dev environments prefer not to install it.
However, this leads to the `verdi-autodocs` hook always failing as it
removes the `verdi tui` section from the docs. The developer then has
to manually reset the changes and run `git commit` with `--no-verify`.

This is annoying enough that as a workaround the `verdi-autodocs` hook
now skips the `verdi tui` command when generating the docs.