Engine: Improve error message when submitting without broker
The `aiida.engine.launch.submit` method was just raising a vague
`AssertionError` in case the runner did not have a communicator, which
is the case if it was constructed without a communicator which in turn
happens for profiles that do not configure a broker.

Since profiles without brokers are now supported and users are bound to
try to submit anyway, the error message should be clearer.