CLI: Add RabbitMQ options to `verdi profile setup` (#6453)
The `verdi profile setup` command was added to replace the deprecated
command `verdi setup`. The new command dynamically generates a
subcommand for each installed storage plugin.

However, the new command did not allow to configure the connection
parameters for the RabbitMQ broker, unlike `verdi setup`. These common
options are now added to the subcommands.

In addition, a new option is added `--use-rabbitmq/--no-use-rabbitmq`.
This flag is on by default, to keep the old behavior of `verdi setup`.
When toggled to `--no-use-rabbitmq`, the RabbitMQ configuration options
are no longer required and are also not prompted for. The profile is
then configured without a broker.