CLI: Remove the RabbitMQ options from `verdi profile setup`
For the vast majority of use cases, users will have a default setup for
RabbitMQ and so the default configuration will be adequate and so they
will not need the options in the command. On the flipside, showing the
options by default can makes the command harder to use as users will
take pause to think what value to pass.

Since there is the `verdi profile configure-rabbitmq` command now that
allows to configure or reconfigure the RabbitMQ connection parameters
for an existing profile, it is fine to remove these options from the
profile setup. Advanced users that need to customize the connection
parameters can resort to that separate command.