👌 CLI: Make `configure-rabbitmq` non-interactive by default
The `configure-rabbitmq` command was introduced mainly to allow new users that set up
their profile with `verdi presto` before they set up RabbitMQ to upgrade their profile
to use the message broker. However, since the command is interactive, they would be
prompted for each of the options when running the command without
`-n`/`--non-interactive`. Users that don't understand these inputs will typically
want to set the defaults, so switching the modus operandi of the commmand to
be non-interactive will make life easier for these users. Users that _do_ want to set
different values than the defaults will understand the options of the command and
should be able to be able to provide them directly or via the interactive mode.